src/test/java/com/bhupi/spring6restmvc/services/BeerCsvServiceImplTest.java
package com.bhupi.spring6restmvc.services;

import com.bhupi.spring6restmvc.model.BeerCSVRecord;
import org.junit.jupiter.api.Test;
import org.springframework.util.ResourceUtils;

import java.io.File;
import java.io.FileNotFoundException;
import java.util.List;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;

class BeerCsvServiceImplTest {

    BeerCsvService beerCsvService = new BeerCsvServiceImpl();

    @Test
    void convertCSV() throws FileNotFoundException {

        File file = ResourceUtils.getFile("classpath:csvdata/beers.csv");

        List<BeerCSVRecord> beerRecords = beerCsvService.convertCSV(file);

        System.out.println(beerRecords.size());

        assertThat(beerRecords.size()).isGreaterThan(0);

    }
}
